@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 
 const DOMParser = typeof window !== 'undefined' && window.DOMParser;
@@ -36,9 +37,9 @@ export default class InlineSVG extends React.Component {
         src: ''
     };
     static propTypes = {
-        src: React.PropTypes.string.isRequired,
-        element: React.PropTypes.string,
-        raw: React.PropTypes.bool
+        src: PropTypes.string.isRequired,
+        element: PropTypes.string,
+        raw: PropTypes.bool
     };
 
     constructor(props) {
